+++ api/data_reader.py
@@ -1,0 +1,24 @@
+import gcsfs
+import pandas as pd
+
+
+def get_links():
+    gcs_url = "gs://movie_recommendation_1/links.csv"
+    return pd.read_csv(gcs_url)
+
+def get_data():
+    gcs_url = "gs://movie_recommendation_1/movies_metadata_updated.csv"
+    return pd.read_csv(gcs_url)
+
+
+def get_matrix():
+    gcs_url = "gs://movie_recommendation_1/matrix_test.csv"
+    return pd.read_csv(gcs_url)
+
+
+def get_credits():
+    gcs_url = "gs://movie_recommendation_1/credits.csv"
+    return pd.read_csv(gcs_url)
+
+
+
