+++ app/static/proposition.html
