+++ app/static/proposition_js.js
