+++ app/static/style_proposition.css
